US236-Task260 updated model test
